Redact contextual clues in number answers
For number answers, the redaction algorithm in the question only redacts
the numeric values, leaving contextual values like the percent sign to
make it easier to answer the question. However, the same logic was not
applied to the correct answer which means that we can guess the correct
answer too easily. For example, we may have a question like "the
atmosphere is made up of ?% CO2" and one of the answers is "52%". That's
too easy! After this change, the correct answer in this example would be
redacted to "52" making the question less easy to answer.