Consider number separators in redaction algorithm
To give enough context for the answer, the redaction algorithm does not
redact context words. However, this sometimes leads to nothing at all
being redacted, e.g. for the question "an Earth year is about 365.26
years long", the correct answer "365.26" will be shortened to "36526"
and then we'll fail to redact that token in the question text.

This patch white-lists comma and dot so that we cover a lot of the
common failure cases like "1,000,000" or "3.14"

Fixes #9